# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateStorageNodeRequestDto": {
        "properties": {
          "attributes": { "additionalProperties": true, "type": "object" },
          "code": { "maxLength": 128, "nullable": true, "type": "string" },
          "nodeType": {
            "enum": ["container", "custom", "rack", "room", "shelf", "site", "zone"],
            "type": "string"
          },
          "parentPublicId": { "format": "uuid", "nullable": true, "type": "string" },
          "title": { "maxLength": 500, "minLength": 1, "type": "string" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["nodeType", "title"],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "detail": { "example": "One or more fields are invalid.", "type": "string" },
          "fieldErrors": {
            "items": { "$ref": "#/components/schemas/ProblemFieldErrorDto" },
            "type": "array"
          },
          "requestId": {
            "example": "0198f40c-92f3-7a12-bc9a-653f97786c2b",
            "format": "uuid",
            "type": "string"
          },
          "status": { "example": 400, "maximum": 599, "minimum": 400, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://inventory-atlas.local/problems/validation-failed",
            "type": "string"
          }
        },
        "required": ["code", "detail", "requestId", "status", "title", "type"],
        "type": "object"
      },
      "ProblemFieldErrorDto": {
        "properties": {
          "field": { "example": "displayName", "type": "string" },
          "messages": {
            "example": ["must contain at least 1 character"],
            "items": { "type": "string" },
            "type": "array"
          }
        },
        "required": ["field", "messages"],
        "type": "object"
      },
...
      "StorageBreadcrumbDto": {
        "properties": {
          "depth": { "minimum": 0, "type": "number" },
          "publicId": { "format": "uuid", "type": "string" },
          "title": { "type": "string" }
        },
        "required": ["depth", "publicId", "title"],
        "type": "object"
      },
      "StorageContentDto": {
        "properties": {
          "kind": { "enum": ["item", "node"], "type": "string" },
          "publicId": { "format": "uuid", "type": "string" },
          "title": { "type": "string" },
          "version": { "minimum": 1, "type": "number" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["kind", "publicId", "title", "version", "visibility"],
        "type": "object"
      },
      "StorageContentPageDto": {
        "properties": {
          "entries": {
            "items": { "$ref": "#/components/schemas/StorageContentDto" },
            "type": "array"
          },
          "nextCursor": { "nullable": true, "type": "string" }
        },
        "required": ["entries"],
        "type": "object"
      },
      "StorageNodeDetailDto": {
        "properties": {
          "attributes": { "additionalProperties": true, "type": "object" },
          "breadcrumb": {
            "items": { "$ref": "#/components/schemas/StorageBreadcrumbDto" },
            "type": "array"
          },
          "code": { "nullable": true, "type": "string" },
          "contents": { "$ref": "#/components/schemas/StorageContentPageDto" },
          "depth": { "minimum": 0, "type": "number" },
          "nodeType": {
            "enum": ["container", "custom", "rack", "room", "shelf", "site", "zone"],
            "type": "string"
          },
          "parentPublicId": { "format": "uuid", "nullable": true, "type": "string" },
          "publicId": { "format": "uuid", "type": "string" },
          "title": { "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "version": { "minimum": 1, "type": "number" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": [
          "attributes",
          "breadcrumb",
          "contents",
          "depth",
          "nodeType",
          "publicId",
          "title",
          "updatedAt",
          "version",
          "visibility"
        ],
        "type": "object"
      },
      "StorageNodePageDto": {
        "properties": {
          "entries": {
            "items": { "$ref": "#/components/schemas/StorageNodeSummaryDto" },
            "type": "array"
          },
          "nextCursor": { "nullable": true, "type": "string" }
        },
        "required": ["entries"],
        "type": "object"
      },
      "StorageNodeSummaryDto": {
        "properties": {
          "code": { "nullable": true, "type": "string" },
          "depth": { "minimum": 0, "type": "number" },
          "nodeType": {
            "enum": ["container", "custom", "rack", "room", "shelf", "site", "zone"],
            "type": "string"
          },
          "parentPublicId": { "format": "uuid", "nullable": true, "type": "string" },
          "publicId": { "format": "uuid", "type": "string" },
          "title": { "type": "string" },
          "version": { "minimum": 1, "type": "number" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["depth", "nodeType", "publicId", "title", "version", "visibility"],
        "type": "object"
      },
...
      "UpdateStorageNodeRequestDto": {
        "properties": {
          "archived": { "type": "boolean" },
          "attributes": { "additionalProperties": true, "type": "object" },
          "code": { "maxLength": 128, "nullable": true, "type": "string" },
          "expectedVersion": { "minimum": 1, "type": "number" },
          "nodeType": {
            "enum": ["container", "custom", "rack", "room", "shelf", "site", "zone"],
            "type": "string"
          },
          "title": { "maxLength": 500, "minLength": 1, "type": "string" },
          "visibility": {
            "enum": ["authenticated", "private", "public", "unlisted"],
            "type": "string"
          }
        },
        "required": ["expectedVersion"],
        "type": "object"
      },
...
      "VersionConflictProblemDto": {
        "properties": {
          "code": { "example": "VALIDATION_FAILED", "type": "string" },
          "currentVersion": { "example": 8, "minimum": 1, "type": "number" },
          "detail": { "example": "One or more fields are invalid.", "type": "string" },
          "fieldErrors": {
            "items": { "$ref": "#/components/schemas/ProblemFieldErrorDto" },
            "type": "array"
          },
          "requestId": {
            "example": "0198f40c-92f3-7a12-bc9a-653f97786c2b",
            "format": "uuid",
            "type": "string"
          },
          "safeDiff": {
            "additionalProperties": true,
            "description": "Visibility-safe fields that differ between the stored Item and the rejected submission. Each entry carries the current and submitted value; fields the actor cannot view are never included.",
            "example": {
              "displayName": {
                "current": "Cordless drill (workshop)",
                "submitted": "Cordless drill"
              }
            },
            "type": "object"
          },
          "status": { "example": 400, "maximum": 599, "minimum": 400, "type": "number" },
          "title": { "example": "Request validation failed", "type": "string" },
          "type": {
            "example": "https://inventory-atlas.local/problems/validation-failed",
            "type": "string"
          }
        },
        "required": [
          "code",
          "currentVersion",
          "detail",
          "requestId",
          "safeDiff",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "cookie": { "in": "cookie", "name": "inventory_atlas_session", "type": "apiKey" }
    }
  },
  "info": {
    "contact": {},
    "description": "The REST contract consumed by Inventory Atlas clients.",
    "title": "Inventory Atlas API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/storage-nodes": {
      "get": {
        "operationId": "listStorageNodes",
        "parameters": [
          { "in": "query", "name": "cursor", "required": false, "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "maximum": 100, "minimum": 1, "type": "number" }
          },
          {
            "in": "query",
            "name": "parentPublicId",
            "required": false,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StorageNodePageDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["storage"]
      },
      "post": {
        "operationId": "createStorageNode",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateStorageNodeRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StorageNodeSummaryDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["storage"]
      }
    },
    "/api/v1/storage-nodes/{publicId}": {
      "get": {
        "operationId": "getStorageNode",
        "parameters": [
          { "in": "query", "name": "cursor", "required": false, "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "maximum": 100, "minimum": 1, "type": "number" }
          },
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StorageNodeDetailDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["storage"]
      },
      "patch": {
        "operationId": "updateStorageNode",
        "parameters": [
          { "in": "header", "name": "If-Match", "required": false, "schema": { "type": "string" } },
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateStorageNodeRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StorageNodeSummaryDto" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" } }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VersionConflictProblemDto" }
              }
            },
            "description": ""
          }
        },
        "security": [{ "cookie": [] }],
        "summary": "",
        "tags": ["storage"]
      }
    },
...
  },
  "servers": [],
...
}